K8S controller to re-launch pods only X times max
Signed-off-by: techcobweb <77053+techcobweb@users.noreply.github.com>

The K8s Controller will launch lots of pods.

If an existing pod is already named the same, then a second attempt to
launch the pod is made with the `-1` suffix.

If that one already exists, then a `-2` suffix is tried. ...etc.

When the limit is reached, the re-launching loop is abandoned, and the
pod is marked as finished with EnvFail.

The retry count is managed using a helm property `maxTestPodRetryLimit`
which defaults to 5.

That gets put into the settings config map, where it is called
`max_test_pod_retry_limit`, so people with kube admin rights can change
this value on the fly, and it will get used within the next 20 seconds.

See issue: https://github.com/galasa-dev/projectmanagement/issues/2263